patch for PullTransaction

+++ src/CAMPServerWebsocketSession/Namespaces/CAMP.Transaction.ts
@@ -158,6 +158,7 @@ export class CAMPTransactionManager extends EventEmitter implements CAMPTransact
         public byteLength = 0n;
         private path: string = "";
         private handle: FileHandle | null = null;
+        private _disposed = false;
 
         constructor(
             private source: Readable,
@@ -169,15 +170,15 @@ export class CAMPTransactionManager extends EventEmitter implements CAMPTransact
          * Buffers the stream once on disk in a temporary directory
          * */
         public async setup() {
-            const pat = path.join(tmpdir(), `CAMP-TX-${this.txId}`);
-            this.path = pat;
+            this.path = path.join(tmpdir(), `CAMP-TX-${process.pid}-${this.txId}-${crypto.randomUUID()}`);
+            ;
 
             //write source to path
-            const writeStream = createWriteStream(pat);
+            const writeStream = createWriteStream(this.path);
             this.source.pipe(writeStream, {end: true});
             await finished(writeStream);
 
-            const stats = await stat(pat, {bigint: true});
+            const stats = await stat(this.path, {bigint: true});
             this.byteLength = stats.size;
 
             this.handle = await open(this.path, "r");
@@ -189,9 +190,13 @@ export class CAMPTransactionManager extends EventEmitter implements CAMPTransact
         }
 
         public async dispose() {
+            if (this._disposed)
+                return;
+
             Guard.AgainstNullish(this.handle);
             await this.handle.close();
             await unlink(this.path);
+            this._disposed = true;
         }
     }
